requirements.txt
streamlit
google-generativeai
python-dotenv
sentence-transformers
PyPDF2
docx2txt
